refactor code and add support for space creation
